test(server): Co-locate metrics body lifecycle tests
Keep response-body outcome coverage next to the MetricsBody state machine while retaining middleware timing coverage with its middleware tests.
